requirements.txt
requests
beautifulsoup4
python-dotenv
groq
gtts
deep-translator
streamlit
pandas
uuid
fastapi
uvicorn
